refactor: move resolvers to dedicated package
