+++ mycore-base/src/main/java/org/mycore/common/xsl/uriresolver/MCRPropertiesResolver.java
@@ -16,7 +16,7 @@
  * along with MyCoRe.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.mycore.common.xml;
+package org.mycore.common.xsl.uriresolver;
 
 import javax.xml.transform.Source;
 import javax.xml.transform.URIResolver;
