+++ mycore-base/src/main/resources/config/mycore.properties
@@ -76,9 +76,9 @@ SAXON=net.sf.saxon.TransformerFactoryImpl
 # Sortable flag for Classification with ID {MCRObjectID}
 # MCR.URIResolver.Classification.Sort.{MCRObjectID}=true
 
-  MCR.URIResolver.ModuleResolver.property=org.mycore.common.xml.MCRPropertiesResolver
-  MCR.URIResolver.ModuleResolver.userobjectrights=org.mycore.common.xml.MCRUserAndObjectRightsURIResolver
-  MCR.URIResolver.ModuleResolver.callJava=org.mycore.common.xml.MCRFunctionResolver
+  MCR.URIResolver.ModuleResolver.property=org.mycore.common.xsl.uriresolver.MCRPropertiesResolver
+  MCR.URIResolver.ModuleResolver.userobjectrights=org.mycore.common.xsl.uriresolver.MCRUserAndObjectRightsURIResolver
+  MCR.URIResolver.ModuleResolver.callJava=org.mycore.common.xsl.uriresolver.MCRFunctionResolver
 
 # Caching URIResolver: cache:...
 
@@ -769,7 +769,7 @@ MCR.Category.LinkService.Class=org.mycore.datamodel.classifications2.impl.MCRCat
 ##############################################################################
 # Rate-Limit-Resolver
 ##############################################################################
-MCR.URIResolver.ModuleResolver.ratelimit=org.mycore.common.xml.MCRRateLimitResolver
+MCR.URIResolver.ModuleResolver.ratelimit=org.mycore.common.xsl.uriresolver.MCRRateLimitResolver
 
 ##############################################################################
 # LegalEntityService
