+++ mycore-base/src/test/java/org/mycore/common/xsl/uriresolver/MCRMockResolver.java
@@ -16,7 +16,7 @@
  * along with MyCoRe.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.mycore.common.xml;
+package org.mycore.common.xsl.uriresolver;
 
 import java.util.ArrayList;
 import java.util.Collections;
